Fixes for s3-sync action
Split s3-sync.yaml out into dev and stable actions. Dev will run on any branch pull that is not stable.

Added DEST_DIR so the sync uploads to 'deployment/latest' for stable and 'deployment/dev' for dev. Otherwise it would upload to the root directory of the bucket.

Should address #24